+++ src/main/scala/dxWDL/compiler/Top.scala
@@ -263,9 +263,7 @@ object Top {
         // Generate dx:applets and dx:workflow from the IR
         Native.apply(irNs,
                      dxWDLrtId, folder, dxProject, instanceTypeDB, dxObjDir,
-                     cOpt.extras,
-                     cOpt.runtimeDebugLevel,
-                     cOpt.leaveWorkflowsOpen, cOpt.force, cOpt.archive, cOpt.locked, cOpt.verbose)
+                     cOpt)
     }