+++ src/test/scala/dxWDL/ExtrasTest.scala
@@ -306,4 +306,66 @@ class ExtrasTest extends FlatSpec with Matchers {
                 "Add" -> DxRunSpec(None, None, None, Some(DxTimeout(None, None, Some(30)))))
         )
     }
+
+    it should "parse the docker registry section" in {
+        val data =
+            """|{
+               | "docker_registry" : {
+               |   "registry" : "foo.bar.dnanexus.com",
+               |   "username" : "perkins",
+               |   "credentials" : "The Bandersnatch has gotten loose"
+               | }
+               |}
+               |""".stripMargin.parseJson
+
+        val extras = Extras.parse(data, verbose)
+        extras.dockerRegistry should be (
+            Some(DockerRegistry(
+                     "foo.bar.dnanexus.com",
+                     "perkins",
+                     "The Bandersnatch has gotten loose")))
+    }
+
+    it should "recognize errors in docker registry section" in {
+        val data =
+            """|{
+               | "docker_registry" : {
+               |   "registry_my" : "foo.bar.dnanexus.com",
+               |   "username" : "perkins",
+               |   "credentials" : "BandersnatchOnTheLoose"
+               | }
+               |}
+               |""".stripMargin.parseJson
+        assertThrows[Exception] {
+            Extras.parse(data, verbose)
+        }
+    }
+
+    it should "recognize errors in docker registry section II" in {
+        val data =
+            """|{
+               | "docker_registry" : {
+               |   "registry" : "foo.bar.dnanexus.com",
+               |   "credentials" : "BandersnatchOnTheLoose"
+               | }
+               |}
+               |""".stripMargin.parseJson
+        assertThrows[Exception] {
+            Extras.parse(data, verbose)
+        }
+    }
+
+
+    it should "recognize errors in docker registry section III" in {
+        val data =
+            """|{
+               | "docker_registry" : {
+               |   "creds" : "XXX"
+               | }
+               |}
+               |""".stripMargin.parseJson
+        assertThrows[Exception] {
+            Extras.parse(data, verbose)
+        }
+    }
 }